+++ RenderingEngine/src/rendering/device/DepthBuffer.cpp
@@ -20,8 +20,8 @@ D3D11_TEXTURE2D_DESC DepthBuffer::CreateDepthTextureDesc(const WindowData& windo
 {
 	D3D11_TEXTURE2D_DESC desc{};
 
-	desc.Width = windowData.screenWidth;
-	desc.Height = windowData.screenHeight;
+	desc.Width = static_cast<UINT>(windowData.screenWidth);
+	desc.Height = static_cast<UINT>(windowData.screenHeight);
 	desc.MipLevels = 1;
 	desc.ArraySize = 1;
 	desc.Format = DXGI_FORMAT_D24_UNORM_S8_UINT;
